moving a few things around- `lap` -> `raft::solver`, `mst`-> `raft::sparse::solver, `single_linkage` -> raft::cluster.
