+++ cpp/include/raft/solver/detail/d_structs.h
