+++ cpp/include/raft/sparse/mst/mst.hpp
@@ -21,4 +21,9 @@
  */
 #pragma once
 
-#include "mst.cuh"
+#pragma message(__FILE__                                                  \
+                " is deprecated and will be removed in a future release." \
+                " Please use the raft/sparse/solver version instead.")
+
+#include <raft/sparse/mst/mst.cuh>
+#include <raft/sparse/mst/mst_solver.cuh>
